+++ client/src/containers/DashboardContainer/DashboardContainer.js
@@ -3,43 +3,28 @@ import Navigation from "../../components/Layout/Navigation/Navigation";
 import Frame from "../../components/Layout/Frame/Frame";
 import Card from "../../components/Layout/Cards/Cards";
 import Resource from "../Resource/Resource";
-import Pagination from "../../components/Layout/Pagination/Pagination";
-import axios from "axios";
-import { getJwt } from "../../helpers/jwt";
+import Pagination from '../LayoutController/PaginationContainer/PaginationContainer';
 
 class DashboardController extends Component {
   constructor(props) {
     super(props);
     this.handler.bind(this);
+    this.handlePagination.bind(this);
   }
 
   state = {
     filter: "all",
-    limiter: 20,
-    displayPagination: false
+    limiter: 20
   };
-
-  async componentDidMount() {
-    try {
-      const data = await axios.get("/api/v1/documents/ping", {
-        headers: { Authorization: `Bearer: ${getJwt()}` }
-      });
-      const pagination = data.data.documents;
-      if (pagination > this.state.limiter) {
-        return this.setState({
-          displayPagination: true
-        });
-      }
-      return this.setState({ displayPagination: false });
-    } catch (error) {
-      console.error(error);
-    }
-  }
 
   handler(e) {
     return this.setState({
       [e.target.name]: String(e.target.textContent).toLocaleLowerCase()
     });
+  }
+
+  handlePagination(e) {
+    return this.setState({ limiter: this.state.limiter + 5 })
   }
 
   render() {
@@ -60,13 +45,7 @@ class DashboardController extends Component {
             }}
           />
         </Frame>
-        {this.state.displayPagination === true ? (
-          <Pagination
-            handlePagination={() =>
-              this.setState({ limiter: this.state.limiter + 5 })
-            }
-          />
-        ) : null}
+        <Pagination handlePagination={e => this.handlePagination(e)} />
       </Fragment>
     );
   }
